requirements.txt
click
coverage
nose
psutil
pylint
python-daemon
rednose

